Fix utils_test after merging
