+++ tests/core_tests/paulis_tests/pauli_test.py
@@ -1127,6 +1127,31 @@ class TestPaulis:
         assert psum1 == psum3
         assert psum2 != psum3
 
+    def test_make_hermitian(self):
+        available_dimensions = [2, 3, 5, 7]
+        # with random phases
+        for i in range(50):
+            n_qubits = np.random.randint(1, 10)
+            dims = [np.random.choice(available_dimensions) for _ in range(n_qubits)]
+            n_paulis = np.random.randint(1, np.min([4**n_qubits - 1, 10]))
+            P = PauliSum.from_random(n_paulis=n_paulis,
+                                     dimensions=dims,
+                                     rand_weights=True)
+            P.set_phases(np.random.randint(0, 2 * P.lcm, n_paulis))
+            P1 = P.make_hermitian()
+            assert P1.is_hermitian()
+
+        # without random phases
+        for i in range(50):
+            n_qubits = np.random.randint(1, 10)
+            dims = [np.random.choice(available_dimensions) for _ in range(n_qubits)]
+            n_paulis = np.random.randint(1, np.min([4**n_qubits - 1, 10]))
+            P = PauliSum.from_random(n_paulis=n_paulis,
+                                     dimensions=dims,
+                                     rand_weights=True)
+            P1 = P.make_hermitian()
+            assert P1.is_hermitian()
+
     def test_ordered_eigenspectrum(self):
         for _ in range(N_tests):
             dimensions = [2, 3, 5, 7]
